requirements.txt
asgiref==3.3.1
Authlib==0.15.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
crypto-cookie==0.3.1
cryptography==3.2.1
Django==3.1.4
idna==2.10
ndg-httpsclient==0.5.1
-e git+https://github.com/cedadev/ndg_saml.git@540186847ade854275bec4e2330ebef69a61050e#egg=ndg_saml
pyasn1==0.4.8
pycparser==2.20
pyOpenSSL==20.0.0
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0